requirements.txt
tox>=2.9.1
coverage>=4.5.1
robotframework==3.0.4
redis==2.10.5
deco>=0.5.2
